requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
email-validator>=2.0,<3.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
httpx
